+++ src/Gsemac.IO/Compression/Implementations/SystemIOCompressionZipArchive.cs
@@ -25,7 +25,7 @@ namespace Gsemac.IO.Compression.Implementations {
             this(stream, true, fileAccess, options) {
         }
 
-        public override IArchiveEntry AddEntry(Stream stream, string entryName, bool leaveOpen = false) {
+        public override IArchiveEntry AddEntry(Stream stream, string entryName, bool overwrite = true, bool leaveOpen = false, IArchiveEntryOptions options = null) {
 
             if (stream is null)
                 throw new ArgumentNullException(nameof(stream));
@@ -38,8 +38,14 @@ namespace Gsemac.IO.Compression.Implementations {
 
             IArchiveEntry existingEntry = GetEntry(entryName);
 
-            if (!(existingEntry is null))
-                DeleteEntry(existingEntry);
+            if (!(existingEntry is null)) {
+
+                if (overwrite)
+                    DeleteEntry(existingEntry);
+                else
+                    throw ArchiveExceptions.EntryAlreadyExists;
+
+            }
 
             System.IO.Compression.ZipArchiveEntry entry = archive.CreateEntry(entryName, GetCompressionLevel(CompressionLevel));
