+++ src/Gsemac.IO/Compression/Implementations/ZipStorerZipArchive.cs
@@ -32,7 +32,7 @@ namespace Gsemac.IO.Compression.Implementations {
             this(stream, leaveOpen: true, fileAccess, options) {
         }
 
-        public override IArchiveEntry AddEntry(Stream stream, string entryName, bool leaveOpen = false) {
+        public override IArchiveEntry AddEntry(Stream stream, string entryName, bool overwrite = true, bool leaveOpen = false, IArchiveEntryOptions options = null) {
 
             if (!CanWrite)
                 throw ArchiveExceptions.ArchiveIsReadOnly;
@@ -48,8 +48,14 @@ namespace Gsemac.IO.Compression.Implementations {
 
             IArchiveEntry existingEntry = GetEntry(entryName);
 
-            if (!(existingEntry is null))
-                DeleteEntry(existingEntry);
+            if (!(existingEntry is null)) {
+
+                if (overwrite)
+                    DeleteEntry(existingEntry);
+                else
+                    throw ArchiveExceptions.EntryAlreadyExists;
+
+            }
 
             ZipStorer.ZipFileEntry entry = archive.Value.AddStream(GetCompression(CompressionLevel), entryName, stream, DateTime.Now);
